Update controller state demo
Made it so that the state cubes show up
above the controller model that they're associated with, which helps the visualization connect a bit better. Also made it so that the axis motion is constrained rather than allowing the cube to "walk" around infinitely, and changed button feedback to reflect pressed state rather than toggle. A few minor code style and ease-of-use changes as well.