Make samples use feature requests and update logo
Updated all the samples to use the proper feature requests for their
content (Mostly this just means requiring 'local-floor' in a bunch of
places.) This should be effectively ignored on implementations like
Chrome 76, so it's best to just have it in place for when browser catch
up.

Also added the new logo and updated (most of) the favicons, as well as
removing the big scary banner declaring that the API thse samples use is
unstable, because it pretty much isn't any more.