Adding handling selectbegin and selectend to Input Selection sample.
Recently, found a bug in Oculus Browser with handling selectbegin/select/selectend which I was unable to detect with the existing tests/samples. Now, it should become obvious, if there are issues with these events.
The proper behavior is:
* On selectbegin the cube must enlarge for 25%
* On select it reduces by 50%
* On selectend it enlarges by 25%; if there was no 'select' event, then the final size will stay at 75%.

At the end, the cubes must appear in the same size as they were, if all selectbegin/select/selectend events are dispatched correctly.