Adding proper cleanup of textures.
After D25543281 lands the layers sample will be broken since they do not have proper calls to gl.clear. The reason why it worked before is the call to gl.invalidateFramebuffer, which is now going to be useless.

A bit more details. Now, the textures going to be invalidated and detached from the framebuffer at the end of rAF. User's code must get the texture and attach to a FBO each frame (each frame when the content is going to be updated). Thus the previous call to gl.invalidateFramebuffer at the beginning of the frame has no effect since there is no color attachment yet. The invalidateFramebuffer caused the clear to happen automatically, but since it has no effect anymore we need to add proper gl.clear calls.